+++ config/testdata/config/app/application.yaml
@@ -14,7 +14,7 @@ dubbo:
         interface: org.github.dubbo.HelloService # must be compatible with grpc or dubbo-java
   provider:
     register: true
-    registry: nacos
+    registries: nacos
     services:
       helloService:
         protocol: dubbo
