+++ config/testdata/config/provider/application.yaml
@@ -6,12 +6,12 @@ dubbo:
       address: nacos://127.0.0.1:8848
   provider:
     register: true
-    registry:
+    registries:
      - nacos
      - zk
     services:
       helloService:
         interface: org.dubbo.service.HelloService
-        registry: nacos,zk
+        registries: nacos,zk
       orderService:
         interface: org.dubbo.service.OrderService
