+++ config/testdata/config/provider/registry_application.yaml
@@ -4,10 +4,10 @@ dubbo:
       timeout: 3s
       address: naocs://127.0.0.1:8848
   provider:
-    registry: nacos
+    registries: nacos
     services:
       HelloService:
         interface: org.dubbo.service.HelloService
-        registry: nacos,zk
+        registries: nacos,zk
       OrderService:
         interface: org.dubbo.service.OrderService
