+++ config/testdata/consumer_config_with_configcenter.yml
@@ -7,7 +7,7 @@ config_center:
   address: "127.0.0.1"
 references:
   "UserProvider":
-    registry: "hangzhouzk,shanghaizk"
+    registries: "hangzhouzk,shanghaizk"
     filter: ""
     protocol : "dubbo"
     interface : "com.ikurento.user.UserProvider"
