+++ config/testdata/consumer_config_withoutProtocol.yml
@@ -33,7 +33,7 @@ registries :
 
 references:
   "UserProvider":
-    registry: "hangzhouzk,shanghaizk"
+    registries: "hangzhouzk,shanghaizk"
     filter: ""
     version: "1.0"
     group: "as"
