+++ config/testdata/provider_config.yml
@@ -27,7 +27,7 @@ registries :
 
 services:
   "UserProvider":
-    registry: "hangzhouzk,shanghaizk"
+    registries: "hangzhouzk,shanghaizk"
     filter: ""
     # the name of limiter
     tps.limiter: "default"
